update tanstackQuery
- separate tanstackQuery;
- separate types;
