+++ src/components/CategoryFilter.tsx
@@ -1,22 +1,16 @@
 import React from 'react';
-import { useQuery } from '@tanstack/react-query';
 import { Form } from 'react-bootstrap';
 
-import { fetchCategories } from '../utils/api';
+import { Category } from '../types/Category.type';
+import { useCategories } from '../hooks/tanstackQuery';
 
-interface Category {
-  idCategory: string;
-  strCategory: string;
-  strCategoryThumb: string;
-  strCategoryDescription: string;
-}
 
 interface CategoryFilterProps {
   onCategoryChange: (category: string | null) => void;
 }
 
 export const CategoryFilter: React.FC<CategoryFilterProps> = ({ onCategoryChange }) => {
-  const { data: categories = [], isLoading, error } = useQuery<Category[], Error>({ queryKey: ['categories'], queryFn: fetchCategories });
+  const { data: categories = [], isLoading, error } = useCategories();
 
   const handleCategoryChange = (evt: React.ChangeEvent<any>) => {
     const category = (evt.target as HTMLSelectElement).value || null; // Explicitly cast event.target
@@ -35,7 +29,6 @@ export const CategoryFilter: React.FC<CategoryFilterProps> = ({ onCategoryChange
     <div className='my-4'>
       <h5 className='mb-3 text-primary'>Filter by Category</h5>
       <Form.Group controlId='categoryFilter' className='mb-3'>
-        {/* <Form.Label className='mb-3 text-primary'>Filter by Category</Form.Label> */}
         <Form.Control
           as='select'
           defaultValue=''
@@ -43,7 +36,7 @@ export const CategoryFilter: React.FC<CategoryFilterProps> = ({ onCategoryChange
         >
           <option value=''>All Categories</option>
           {Array.isArray(categories) &&
-            categories.map((category) => (
+            categories.map((category: Category) => (
               <option key={category.idCategory} value={category.strCategory}>
                 {category.strCategory}
               </option>
