+++ src/types/Category.type.ts
@@ -1,0 +1,6 @@
+export interface Category {
+  idCategory: string;
+  strCategory: string;
+  strCategoryThumb: string;
+  strCategoryDescription: string;
+}
