+++ src/types/Recipe.type.ts
@@ -1,0 +1,10 @@
+export interface Recipe {
+  idMeal: string;
+  strMeal: string;
+  strCategory: string;
+  strArea: string;
+  strInstructions: string;
+  strMealThumb: string;
+  [key: `strIngredient${number}`]: string | undefined; // Allow dynamic ingredient keys
+  [key: `strMeasure${number}`]: string | undefined; // Allow dynamic measure keys
+}
